touch ups
* try to get the resource in upgrade()
* separate out slurm managers in to their own separate files
